Switch ShoeboxModuleTest to use ServiceModuleTest and fix a few things: - broken dev dependencies (Zookeeper, MailToKeepPlugin) - AboutAssets Controller - Remove Awaits in AdminLDAController
