+++ kifi-backend/modules/shoebox/app/com/keepit/admin/AdminLDAController.scala
@@ -4,11 +4,9 @@ import com.google.inject.Inject
 import com.keepit.common.controller.{ ActionAuthenticator, AdminController }
 import com.keepit.cortex.CortexServiceClient
 import com.keepit.shoebox.ShoeboxServiceClient
-import views.html
 import play.api.libs.json._
-import scala.concurrent.Await
-import scala.concurrent.duration._
-import com.keepit.cortex.models.lda.{ LDAUserURIInterestScores, LDATopicConfigurations, LDATopicDetail, LDATopicConfiguration }
+import scala.concurrent.Future
+import com.keepit.cortex.models.lda.{ LDATopicDetail, LDATopicConfiguration }
 import com.keepit.common.db.slick.Database
 import com.keepit.model.{ NormalizedURIRepo, NormalizedURI, KeepRepo, User }
 import com.keepit.common.db.Id
@@ -23,13 +21,12 @@ class AdminLDAController @Inject() (
     uriRepo: NormalizedURIRepo,
     keepRepo: KeepRepo) extends AdminController(actionAuthenticator) {
 
-  var ldaConf = Await.result(cortex.ldaConfigurations, 5 seconds)
-
   val MAX_WIDTH = 15
 
-  def index() = AdminHtmlAction.authenticated { implicit request =>
-    val n = Await.result(cortex.ldaNumOfTopics, 5 seconds)
-    Ok(html.admin.lda(n))
+  def index() = AdminHtmlAction.authenticatedAsync { implicit request =>
+    cortex.ldaNumOfTopics.map { n =>
+      Ok(html.admin.lda(n))
+    }
   }
 
   private def trimLongString(s: String) = {
@@ -44,43 +41,39 @@ class AdminLDAController @Inject() (
     }.mkString("\n")
   }
 
-  private def topicName(topicId: Int): String = ldaConf.configs(topicId.toString).topicName
-
-  def showTopics() = AdminHtmlAction.authenticated { implicit request =>
+  def showTopics() = AdminHtmlAction.authenticatedAsync { implicit request =>
     val body = request.body.asFormUrlEncoded.get.mapValues(_.head)
     val fromId = body.get("fromId").get.toInt
     val toId = body.get("toId").get.toInt
     val topN = body.get("topN").get.toInt
-    val res = Await.result(cortex.ldaShowTopics(fromId, toId, topN), 5 seconds)
-
-    val ids = res.map { _.topicId }
-    val topics = res.map { x => getFormatted(x.topicWords) }
-    val names = res.map { _.config.topicName }
-    val states = res.map { _.config.isActive }
-    val js = Json.obj("ids" -> ids, "topicWords" -> topics, "topicNames" -> names, "states" -> states)
-
-    Ok(js)
+    cortex.ldaShowTopics(fromId, toId, topN).map { res =>
+      val ids = res.map { _.topicId }
+      val topics = res.map { x => getFormatted(x.topicWords) }
+      val names = res.map { _.config.topicName }
+      val states = res.map { _.config.isActive }
+      val js = Json.obj("ids" -> ids, "topicWords" -> topics, "topicNames" -> names, "states" -> states)
+      Ok(js)
+    }
   }
 
-  private def showTopTopicDistributions(arr: Array[Float], topK: Int = 5): String = {
-    arr.zipWithIndex.sortBy(-1f * _._1).take(topK).map {
-      case (score, topicId) =>
-        val tname = topicName(topicId)
-        s"$topicId ($tname) : " + "%.3f".format(score)
-    }.mkString(", ")
+  private def showTopTopicDistributions(arr: Array[Float], topK: Int = 5): Future[String] = {
+    cortex.ldaConfigurations.map { ldaConf =>
+      arr.zipWithIndex.sortBy(-1f * _._1).take(topK).map {
+        case (score, topicId) =>
+          val tname = ldaConf.configs(topicId.toString).topicName
+          s"$topicId ($tname) : " + "%.3f".format(score)
+      }.mkString(", ")
+    }
   }
 
-  def wordTopic() = AdminHtmlAction.authenticated { implicit request =>
+  def wordTopic() = AdminHtmlAction.authenticatedAsync { implicit request =>
     val body = request.body.asFormUrlEncoded.get.mapValues(_.head)
     val word = body.get("word").get
-    val res = Await.result(cortex.ldaWordTopic(word), 5 seconds)
-
-    val msg = res match {
+    val futureMsg = cortex.ldaWordTopic(word).flatMap {
       case Some(arr) => showTopTopicDistributions(arr)
-      case None => "word not in dictionary"
+      case None => Future.successful("word not in dictionary")
     }
-
-    Ok(msg)
+    futureMsg.map(msg => Ok(JsString(msg)))
   }
 
   def saveEdits() = AdminHtmlAction.authenticated(parse.tolerantJson) { implicit request =>
@@ -91,22 +84,18 @@ class AdminLDAController @Inject() (
     val isActive = (js \ "topic_states").as[JsArray].value.map { _.as[Boolean] }
 
     val config = (ids, names, isActive).zipped.map { case (id, name, active) => id -> LDATopicConfiguration(name, active) }.toMap
-    ldaConf = LDATopicConfigurations(ldaConf.configs ++ config)
     cortex.saveEdits(config)
     Ok
   }
 
-  def docTopic() = AdminHtmlAction.authenticated { implicit request =>
+  def docTopic() = AdminHtmlAction.authenticatedAsync { implicit request =>
     val body = request.body.asFormUrlEncoded.get.mapValues(_.head)
     val doc = body.get("doc").get
-    val res = Await.result(cortex.ldaDocTopic(doc), 5 seconds)
-
-    val msg = res match {
+    val futureMsg = cortex.ldaDocTopic(doc).flatMap {
       case Some(arr) => showTopTopicDistributions(arr)
-      case None => "not enough information."
+      case None => Future.successful("not enough information.")
     }
-
-    Ok(msg)
+    futureMsg.map(msg => Ok(JsString(msg)))
   }
 
   def userTopicDump(userId: Id[User], limit: Int) = AdminHtmlAction.authenticatedAsync { implicit request =>
@@ -133,13 +122,11 @@ class AdminLDAController @Inject() (
     val body = request.body.asFormUrlEncoded.get.mapValues(_.head)
     val userId = body.get("userId").get.toLong
 
-    cortex.userTopicMean(Id[User](userId)).map { meanOpt =>
-      val res = meanOpt.map { arr =>
-        showTopTopicDistributions(arr, topK = 10)
-      } getOrElse "not enough information"
-      Ok(Json.toJson(res))
+    val futureMsg = cortex.userTopicMean(Id[User](userId)).flatMap {
+      case Some(arr) => showTopTopicDistributions(arr, topK = 10)
+      case None => Future.successful("not enough information")
     }
-
+    futureMsg.map(msg => Ok(JsString(msg)))
   }
 
   def topicDetail(topicId: Int) = AdminHtmlAction.authenticatedAsync { implicit request =>
